Rename handlers module to request_handlers.
